fix(gemma): Address PR review comments
- Extracted port and enabled status resolution into a shared `resolveGemmaConfig` utility in `platform.ts` to reduce duplication.
- Removed the default value for the `--port` option in the `status` command so that the workspace settings act as the single source of truth when the flag is omitted.
