+++ packages/cli/src/commands/gemma/start.ts
@@ -21,6 +21,7 @@ import {
   getBinaryPath,
   isBinaryInstalled,
   isServerRunning,
+  resolveGemmaConfig,
 } from './platform.js';
 
 /**
@@ -87,21 +88,8 @@ export const startCommand: CommandModule = {
     }
 
     if (!port) {
-      try {
-        const { loadSettings } = await import('../../config/settings.js');
-        const settings = loadSettings(process.cwd());
-        const hostStr =
-          settings.merged.experimental?.gemmaModelRouter?.classifier?.host;
-        if (hostStr) {
-          const match = hostStr.match(/:(\d+)/);
-          if (match) {
-            port = parseInt(match[1], 10);
-          }
-        }
-      } catch {
-        // Ignore
-      }
-      port = port ?? DEFAULT_PORT;
+      const { configuredPort } = resolveGemmaConfig(DEFAULT_PORT);
+      port = configuredPort;
     }
 
     if (!isBinaryInstalled()) {
